Switch from using ID to a name slug for credits page
IDs in URLs just suck. I mean, they're not great. If we can avoid them
-somehow- it's worth doing. So we'll use name slugs instead. It requires
case sensitivity, which I don't love, and we have to replace spaces with
underscores so that people can have two names with a space between as a
first name (or a last!), but it works.