Add dynamic audition page
Automatically gets all future shows that do not have the
`noLongerAuditioning` flag set to true. This is a checkbox that can be
toggled for a given show in the CMS.

Right now a show needs to have an opening date set that is in the future
to show up in this list.